+++ lib/features/cart/domain/entity/cart_entity.dart
@@ -3,11 +3,11 @@ import 'package:fruit_hub/core/entities/product_entity.dart';
 class CartEntity {
   final ProductEntity productEntity;
   int count;
-  double unit;
+  int unit;
 
   CartEntity({
     required this.productEntity,
     this.count = 0,
-    this.unit = 0.0,
+    this.unit = 0,
   });
 }
